# backend/tests/fixtures/openapi_before_mvc.json
{
  "components": {
    "schemas": {
...
      "Body_confirm_bkash_checkout__payment_id__confirm_post": {
        "properties": {
          "pin": {
            "title": "Pin",
            "type": "string"
          },
          "wallet_number": {
            "title": "Wallet Number",
            "type": "string"
          }
        },
        "required": [
          "wallet_number",
          "pin"
        ],
        "title": "Body_confirm_bkash_checkout__payment_id__confirm_post",
        "type": "object"
      },
...
      "CreateComplaintRequest": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "reported_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Id"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "category",
          "subject",
          "description"
        ],
        "title": "CreateComplaintRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewRequest": {
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "reviewee_id": {
            "title": "Reviewee Id",
            "type": "string"
          },
          "ride_id": {
            "title": "Ride Id",
            "type": "string"
          },
          "stars": {
            "title": "Stars",
            "type": "integer"
          }
        },
        "required": [
          "ride_id",
          "reviewee_id",
          "stars"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
...
      "UpdateComplaintRequest": {
        "properties": {
          "admin_notes": {
            "default": "",
            "title": "Admin Notes",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateComplaintRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Student-to-student ride-sharing network — Backend API",
    "title": "Arooohi API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/complaints/": {
      "get": {
        "description": "Get complaints — own complaints for users, all complaints for admin.",
        "operationId": "get_complaints_api_complaints__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Complaints",
        "tags": [
          "Complaints"
        ]
      },
      "post": {
        "description": "File a new complaint.\n\nImprovement note:\n- This shared admin moderation API is now stricter about complaint content before it reaches the moderator queue.",
        "operationId": "create_complaint_api_complaints__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Complaint",
        "tags": [
          "Complaints"
        ]
      }
    },
...
    "/api/complaints/{complaint_id}": {
      "patch": {
        "description": "Admin: Update complaint status and add notes.",
        "operationId": "update_complaint_api_complaints__complaint_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComplaintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Complaint",
        "tags": [
          "Complaints"
        ]
      }
    },
...
    "/api/drivers/{profile_id}/review": {
      "patch": {
        "description": "Admin: Approve or reject a driver application.",
        "operationId": "review_driver_api_drivers__profile_id__review_patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Driver",
        "tags": [
          "Driver Verification"
        ]
      }
    },
...
    "/api/reviews": {
      "post": {
        "description": "Leave a 1-5 star review with an optional comment, after a completed ride.",
        "operationId": "create_review_api_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review",
        "tags": [
          "Driver Rating & Review"
        ]
      }
    },
    "/api/reviews/driver/{driver_id}": {
      "get": {
        "description": "Public profile rating: average, star histogram and recent comments.",
        "operationId": "driver_reviews_api_reviews_driver__driver_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Driver Reviews",
        "tags": [
          "Driver Rating & Review"
        ]
      }
    },
...
    "/bkash/checkout/{payment_id}": {
      "get": {
        "description": "The gateway's hosted PIN screen.",
        "operationId": "checkout_page_bkash_checkout__payment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "title": "Payment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "default": "",
              "title": "Error",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Checkout Page",
        "tags": [
          "bKash (simulated)"
        ]
      }
    },
    "/bkash/checkout/{payment_id}/cancel": {
      "post": {
        "description": "User backed out on the gateway's page.",
        "operationId": "cancel_bkash_checkout__payment_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "bKash (simulated)"
        ]
      }
    },
    "/bkash/checkout/{payment_id}/confirm": {
      "post": {
        "description": "User authorised on the gateway. Redirect back to the merchant.\n\nNote what this does NOT do: it never credits anything. It only marks the\npayment authorised and bounces the browser home with a status parameter the\nmerchant is expected to distrust and verify.",
        "operationId": "confirm_bkash_checkout__payment_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_confirm_bkash_checkout__payment_id__confirm_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm",
        "tags": [
          "bKash (simulated)"
        ]
      }
    },
...
  }
}